+++ CHANGELOG.md
@@ -3,3 +3,9 @@
 ## [1.0.0] - 2024-02-21
 
  - Initial release
+
+
+## [1.0.1] - 2024-02-22
+
+ - Add track & trace endpoint
+ - Add piece endpoint